+++ docs/src/app/(doclike)/(default)/playground/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 
-// import { StandaloneEditor } from '@/components/editor/standalone-editor';
+import { StandaloneEditor } from '@/components/editor/standalone-editor';
 
 export const metadata: Metadata = {
   title: 'Playground',
@@ -11,7 +11,7 @@ export default async function Playground() {
     <>
       <h1 className='mb-4'>Grit Playground</h1>
       <div className='h-[80vh]'>
-        Sorry, the playground is currently broken.
+        <StandaloneEditor />
       </div>
     </>
   );
